backend-ai/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1

langchain==0.1.14
langchain-core==0.1.53
langsmith==0.1.17
langchain-text-splitters==0.0.2
langchain-google-genai==0.0.8
google-generativeai==0.3.2
